fix: correct uint16 flag casting in cligen using flagReturnType
The cobra flag getter for uint16 returns int32 (no native uint16 flag).
Added flagReturnType() to determine the actual Go type returned by each
flag getter, and cast to the param type when they differ.

Verified on device: thermal, trace, egl info all work correctly.

+++ cmd/ndkcli/font_gen.go
@@ -75,7 +75,7 @@ var fontMatcherSetStyleCmd = &cobra.Command{
 		defer obj.Close()
 		weight, _ := cmd.Flags().GetInt32("weight")
 		italic, _ := cmd.Flags().GetBool("italic")
-		result := obj.SetStyle(weight, italic)
+		result := obj.SetStyle(uint16(weight), italic)
 		fmt.Println(result)
 		return nil
 	},
